Remove duplicates in the navbar components and tidy things a little bit
