+++ Vouchers/client/src/usersView/NavbarNegocio.js
@@ -4,20 +4,22 @@ import ListItem from '@material-ui/core/ListItem';
 import ListItemIcon from '@material-ui/core/ListItemIcon';
 import ListItemText from '@material-ui/core/ListItemText';
 import InboxIcon from '@material-ui/icons/Inbox';
+import navegacion from '../utils/navegacion';
+import constantes from "../utils/constantes";
 
 
 export default function NavbarNegocio() {
-    
-    const redireccionarAPerfil = () => {
-        window.location.replace('/negocios/perfil');
-    };
+
+  const redirigirAMiNegocio = () => {
+    window.location.replace(navegacion.getNegocioPerfilUrl());
+  };
 
   return (
     <div>
-        <ListItem button key={"Mi Negocio"} onClick={redireccionarAPerfil}>
-            <ListItemIcon> <InboxIcon /> </ListItemIcon>
-            <ListItemText primary={"Mi Negocio"} />
-        </ListItem>
+      <ListItem button key={constantes.miNegocioTitulo} onClick={redirigirAMiNegocio}>
+        <ListItemIcon> <InboxIcon /> </ListItemIcon>
+        <ListItemText primary={constantes.miNegocioTitulo} />
+      </ListItem>
     </div>
   );
 }
