+++ Vouchers/client/src/utils/constantes.js
@@ -1,6 +1,7 @@
 
 export default {
   miPerfilTitulo: 'Mi Perfil',
+  miNegocioTitulo: 'Mi Negocio',
   comprarVouchersTitulo: 'Comprar Vouchers',
   canjearVouchersTitulo: 'Canjear Vouchers',
   reclamosTitulo: 'Reclamos',
